// components/settings/shared/sales-table/models/conversationColumnMeta.ts
export enum ConversationColumnHeader {
  CUSTOMER = 'Customer',
  AGENTS = 'Agents',
  AI_SCORE = 'AI Score',
  CUSTOMER_SENTIMENT = 'Call Sentiment',
  INTENT = 'Intent',
  VEHICLE = 'Vehicle',
  OUTCOME = 'Outcome',
  DATE_AND_TIME = 'Date & Time',
  DURATION = 'Duration',
  ACTION_ITEMS = 'Action Items',
  NOTES = 'Notes',
}

export enum ConversationColumnID {
  CUSTOMER = '1',
  AGENTS = '2',
  AI_SCORE = '3',
  CUSTOMER_SENTIMENT = '4',
  INTENT = '5',
  VEHICLE = '6',
  OUTCOME = '7',
  DATE_AND_TIME = '8',
  DURATION = '9',
  ACTION_ITEMS = '10',
  NOTES = '11',
}